Fix a few compiler warnings

+++ SourceX/storm/storm.cpp
@@ -157,7 +157,7 @@ BOOL SFileOpenFile(const char *filename, HANDLE *phFile)
 
 	if (directFileAccess) {
 		char directPath[DVL_MAX_PATH] = "\0";
-		for (int i = 0; i < strlen(filename); i++) {
+		for (size_t i = 0; i < strlen(filename); i++) {
 			directPath[i] = AsciiToLowerTable_Path[filename[i]];
 		}
 		result = SFileOpenFileEx((HANDLE)0, directPath, 0xFFFFFFFF, phFile);
